requirements.txt
tree_sitter_python
tree_sitter
google-genai
pydantic-settings
